Add `#[inline]` and `#[cold]` where useful to split out the cold path (#7)
LLVM seems to have a hard time properly inlining the various utility
methods in the parser. By strategically adding `#[inline]` and `#[cold]`
annotations and splitting out the hot and cold path this PR makes so
that llvm is able to inline almost the entirety of `parse_record_impl`
and reap the resulting optimizations.