Simplify how `Read` and `read_value` types are exposed (#19)
As it turns out, PERF_RECORD_READ events are only emitted when the
inherit and inherit_stat flags are set. inherit is incompatible with
READ_FORMAT_GROUP so PERF_RECORD_READ will never be emitted with a group
format. This PR takes advantage of that to simplify some things and I've
also gone and rethought how `ReadGroup` and `ReadValue` should be used.

- Sample now always contains a ReadGroup and converts if the source
read_format is only for a single value.
- Read now only contains a ReadValue since it is not possible to create
a Read record when GROUP is set in read_format
- ReadGroup can now be converted into ReadValue via TryFrom
- ReadValue can now be converted into ReadGroup via From
- ReadData has been deleted since the conversions take its place