Fix an unbounded memory usage bug within ParseBufCursor::new (#32)
The symptom was that parsing would work fine for a while and then all of
a sudden the program would start using GBs worth of memory and get OOM
killed. The root cause was that ParseBufCursor was never calling
ParseBuf::advance.

This didn't have any effects most of the time due to the way the code
was structured. However, if the source buffer started off with an empty
chunk then that would be enough to enter an infinite loop.

The fix is to have ParseBufCursor::new properly call ParseBuf::advance
and to only insert chunks into the vector if they have a non-zero
length. The second part isn't required but it ensures any such bugs in
the future (i.e. on the part of a ParseBuf implementation) result in an
infinite loop instead of OOM.